R Released Server v0.3.5.1 + Bug: repodding fails with a bad index error - Removed unnecessary @Overrides

+++ trunk/src/server/campaign/operations/ShortResolver.java
@@ -2564,10 +2564,10 @@ public class ShortResolver {
                         int conquestUnitAdjust = o.getIntValue("AttackerConquestUnitAdjustment");
                         int conquestBVAdjust = o.getIntValue("AttackerConquestBVAdjustment");
                         if (conquestUnitAdjust > 0) {
-                            totalConquest += Math.floor(so.getStartingUnits() / conquestUnitAdjust);
+                            totalConquest += (int) (Math.floor(so.getStartingUnits() / conquestUnitAdjust));
                         }
                         if (conquestBVAdjust > 0) {
-                            totalConquest += Math.floor(so.getStartingBV() / conquestBVAdjust);
+                            totalConquest += (int) (Math.floor(so.getStartingBV() / conquestBVAdjust));
                         }
                         CampaignData.mwlog.testLog("Old Land: " + totalConquest);
                         if (modifyBasedOnPosition) {
